Parameter match

+++ src/motor_listener.py
@@ -5,7 +5,9 @@ import json
 import time
 
 
-def run_motors(drive_left: int, drive_right: int, run_time: float, max_power: float = 0.8):
+def run_motors(
+    drive_left: int, drive_right: int, run_time: float, max_power: float = 0.8
+):
     TB.set_motor_power(1, drive_right * max_power)
     TB.set_motor_power(2, drive_left * max_power)
     time.sleep(run_time)
@@ -13,7 +15,7 @@ def run_motors(drive_left: int, drive_right: int, run_time: float, max_power: fl
 
 
 def move(direction: int, speed: float, time: float):
-    motor_speed = direction * speed
+    motor_speed = int(direction * speed)
     run_motors(motor_speed, motor_speed, time)
